requirements.txt
matplotlib==3.10.7
numpy==2.3.4
opencv_contrib_python==4.12.0.88
opencv_python==4.12.0.88
opencv_python_headless==4.12.0.88
pandas==2.3.3
Pillow==12.0.0
scikit_learn==1.7.2
seaborn==0.13.2
torch==2.8.0+cu126
torchvision==0.23.0+cu126
tqdm==4.67.1
wandb==0.22.2
